fix: prevent fatal error at init if logs index already present
